worker implementation that ingests storybooks and uploads screenshots (#7)

+++ api/src/index.test.ts
@@ -1,4 +1,5 @@
 import request from "supertest"
+import { expect, describe, it, afterAll } from "vitest"
 
 import { Database } from "./database"
 import server from "./index"
@@ -18,8 +19,18 @@ describe("app", () => {
   })
 
   afterAll(async () => {
-    server.close()
-    const db = await Database()
-    await db.destroy()
+    await new Promise<void>((resolve) => {
+      server.close(() => resolve())
+    })
+
+    try {
+      const db = await Database()
+      if (db.isInitialized) {
+        await db.destroy()
+      }
+    } catch (err) {
+      // Do nothing
+      void err
+    }
   })
 })
